habhub-dataserver/requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==3.0.6 # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3 # https://github.com/antirez/redis
hiredis==1.1.0
celery==4.4.6 # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0
flower==0.9.5  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.0.10  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-import-export==2.4.0
django-storages[boto3]==1.11  # https://github.com/jschneier/django-storages

# Django REST Framework
djangorestframework==3.12.1  # https://github.com/encode/django-rest-framework
djangorestframework-gis==0.15.0 # https://github.com/openwisp/django-rest-framework-gis
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.2.0
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-rest-multiple-models==2.1.3

# Custom requirements
jsonfield==3.1.0
django-geojson==3.0.0
django-leaflet==0.27.1
django-summernote==0.8.11.6
geopy==2.0.0
django-multiselectfield==0.1.12
requests==2.24.0
